+++ gno.land/cmd/gnoland/genesis.go
@@ -1,25 +1,17 @@
 package main
 
 import (
-	"context"
 	"flag"
-	"os"
 
 	"github.com/gnolang/gno/tm2/pkg/commands"
 )
 
-func main() {
-	io := commands.NewDefaultIO()
-	cmd := newRootCmd(io)
-
-	cmd.Execute(context.Background(), os.Args[1:])
-}
-
-func newRootCmd(io commands.IO) *commands.Command {
+func newGenesisCmd(io commands.IO) *commands.Command {
 	cmd := commands.NewCommand(
 		commands.Metadata{
-			ShortUsage: "<subcommand> [flags] [<arg>...]",
-			LongHelp:   "Gno Genesis manipulation suite",
+			Name:       "genesis",
+			ShortUsage: "genesis <subcommand> [flags] [<arg>...]",
+			LongHelp:   "gno genesis manipulation suite",
 		},
 		commands.NewEmptyConfig(),
 		commands.HelpExec,
