+++ gno.land/cmd/gnoland/root.go
@@ -35,6 +35,7 @@ func newRootCmd(io commands.IO) *commands.Command {
 		newStartCmd(io),
 		newSecretsCmd(io),
 		newConfigCmd(io),
+		newGenesisCmd(io),
 	)
 
 	return cmd
