+++ gno.land/cmd/gnoland/validator_remove_test.go
@@ -4,23 +4,24 @@ import (
 	"context"
 	"testing"
 
-	"github.com/gnolang/gno/gno.land/pkg/gnoland"
 	"github.com/gnolang/gno/tm2/pkg/bft/types"
 	"github.com/gnolang/gno/tm2/pkg/commands"
-	"github.com/gnolang/gno/tm2/pkg/std"
 	"github.com/gnolang/gno/tm2/pkg/testutils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestGenesis_Balances_Remove(t *testing.T) {
+func TestGenesis_Validator_Remove(t *testing.T) {
 	t.Parallel()
 
-	t.Run("invalid genesis", func(t *testing.T) {
+	t.Run("invalid genesis file", func(t *testing.T) {
+		t.Parallel()
+
 		// Create the command
 		cmd := newRootCmd(commands.NewTestIO())
 		args := []string{
-			"balances",
+			"genesis",
+			"validator",
 			"remove",
 			"--genesis-path",
 			"dummy-path",
@@ -28,25 +29,96 @@ func TestGenesis_Balances_Remove(t *testing.T) {
 
 		// Run the command
 		cmdErr := cmd.ParseAndRun(context.Background(), args)
-		require.ErrorContains(t, cmdErr, errUnableToLoadGenesis.Error())
+		assert.ErrorContains(t, cmdErr, errUnableToLoadGenesis.Error())
 	})
 
-	t.Run("genesis app state not set", func(t *testing.T) {
+	t.Run("invalid validator address", func(t *testing.T) {
 		t.Parallel()
-
-		dummyKey := getDummyKey(t)
 
 		tempGenesis, cleanup := testutils.NewTestFile(t)
 		t.Cleanup(cleanup)
 
 		genesis := getDefaultGenesis()
-		genesis.AppState = nil // not set
 		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
 
 		// Create the command
 		cmd := newRootCmd(commands.NewTestIO())
 		args := []string{
-			"balances",
+			"genesis",
+			"validator",
+			"remove",
+			"--genesis-path",
+			tempGenesis.Name(),
+			"--address",
+			"dummyaddress",
+		}
+
+		// Run the command
+		cmdErr := cmd.ParseAndRun(context.Background(), args)
+		assert.ErrorContains(t, cmdErr, "invalid validator address")
+	})
+
+	t.Run("validator not found", func(t *testing.T) {
+		t.Parallel()
+
+		tempGenesis, cleanup := testutils.NewTestFile(t)
+		t.Cleanup(cleanup)
+
+		dummyKeys := getDummyKeys(t, 2)
+		genesis := getDefaultGenesis()
+
+		// Set an existing validator
+		genesis.Validators = append(genesis.Validators, types.GenesisValidator{
+			Address: dummyKeys[0].Address(),
+			PubKey:  dummyKeys[0],
+			Power:   1,
+			Name:    "example",
+		})
+
+		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
+
+		// Create the command
+		cmd := newRootCmd(commands.NewTestIO())
+		args := []string{
+			"genesis",
+			"validator",
+			"remove",
+			"--genesis-path",
+			tempGenesis.Name(),
+			"--address",
+			dummyKeys[1].Address().String(),
+		}
+
+		// Run the command
+		cmdErr := cmd.ParseAndRun(context.Background(), args)
+		assert.ErrorContains(t, cmdErr, errValidatorNotPresent.Error())
+	})
+
+	t.Run("validator removed", func(t *testing.T) {
+		t.Parallel()
+
+		tempGenesis, cleanup := testutils.NewTestFile(t)
+		t.Cleanup(cleanup)
+
+		dummyKey := getDummyKey(t)
+
+		genesis := getDefaultGenesis()
+
+		// Set an existing validator
+		genesis.Validators = append(genesis.Validators, types.GenesisValidator{
+			Address: dummyKey.Address(),
+			PubKey:  dummyKey,
+			Power:   1,
+			Name:    "example",
+		})
+
+		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
+
+		// Create the command
+		cmd := newRootCmd(commands.NewTestIO())
+		args := []string{
+			"genesis",
+			"validator",
 			"remove",
 			"--genesis-path",
 			tempGenesis.Name(),
@@ -56,85 +128,6 @@ func TestGenesis_Balances_Remove(t *testing.T) {
 
 		// Run the command
 		cmdErr := cmd.ParseAndRun(context.Background(), args)
-		require.ErrorContains(t, cmdErr, errAppStateNotSet.Error())
-	})
-
-	t.Run("address is present", func(t *testing.T) {
-		t.Parallel()
-
-		dummyKey := getDummyKey(t)
-
-		tempGenesis, cleanup := testutils.NewTestFile(t)
-		t.Cleanup(cleanup)
-
-		genesis := getDefaultGenesis()
-		state := gnoland.GnoGenesisState{
-			// Set an initial balance value
-			Balances: []gnoland.Balance{
-				{
-					Address: dummyKey.Address(),
-					Amount:  std.NewCoins(std.NewCoin("ugnot", 100)),
-				},
-			},
-		}
-		genesis.AppState = state
-		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
-
-		// Create the command
-		cmd := newRootCmd(commands.NewTestIO())
-		args := []string{
-			"balances",
-			"remove",
-			"--genesis-path",
-			tempGenesis.Name(),
-			"--address",
-			dummyKey.Address().String(),
-		}
-
-		// Run the command
-		cmdErr := cmd.ParseAndRun(context.Background(), args)
-		require.NoError(t, cmdErr)
-
-		// Validate the genesis was updated
-		genesis, loadErr := types.GenesisDocFromFile(tempGenesis.Name())
-		require.NoError(t, loadErr)
-
-		require.NotNil(t, genesis.AppState)
-
-		state, ok := genesis.AppState.(gnoland.GnoGenesisState)
-		require.True(t, ok)
-
-		assert.Len(t, state.Balances, 0)
-	})
-
-	t.Run("address not present", func(t *testing.T) {
-		t.Parallel()
-
-		dummyKey := getDummyKey(t)
-
-		tempGenesis, cleanup := testutils.NewTestFile(t)
-		t.Cleanup(cleanup)
-
-		genesis := getDefaultGenesis()
-		state := gnoland.GnoGenesisState{
-			Balances: []gnoland.Balance{}, // Empty initial balance
-		}
-		genesis.AppState = state
-		require.NoError(t, genesis.SaveAs(tempGenesis.Name()))
-
-		// Create the command
-		cmd := newRootCmd(commands.NewTestIO())
-		args := []string{
-			"balances",
-			"remove",
-			"--genesis-path",
-			tempGenesis.Name(),
-			"--address",
-			dummyKey.Address().String(),
-		}
-
-		// Run the command
-		cmdErr := cmd.ParseAndRun(context.Background(), args)
-		require.ErrorContains(t, cmdErr, errBalanceNotFound.Error())
+		assert.NoError(t, cmdErr)
 	})
 }
